+++ src/breakpad/resolver.rs
@@ -183,7 +183,7 @@ impl BreakpadResolver {
             Vec::new()
         };
 
-        sym.code_info = Some(direct_code_info);
+        sym.code_info = Some(Box::new(direct_code_info));
         sym.inlined = inlined.into_boxed_slice();
 
         Ok(())